Edit favorite controller
